+++ src/pages/api/payments/[paymentId]/pay.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
-import { ErrorResponse } from '../../../../api'
-import { Payments } from '../../../data/payment'
+import { ErrorResponse } from '../../ErrorResponse'
+import { Payments } from '../../../data/Payments'
 import { delay } from '../../delay';
 
 export default async function handler(
