+++ src/pages/api/payments/index.ts
@@ -1,6 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
-import { ErrorResponse } from '../../../api'
-import { Payment, Payments } from '../../data/payment'
+import { ErrorResponse } from '../ErrorResponse'
+import { Payment } from '../../data/payment2'
+import { Payments } from '../../data/Payments'
 import { delay } from '../delay';
 
 export default async function handler(
