+++ src/pages/data/PaymentStatus.ts
@@ -1,0 +1,1 @@
+export type PaymentStatus = 'Unpaid' | 'Paid';
