+++ src/pages/data/payment2.ts
@@ -1,0 +1,10 @@
+import { PaymentStatus } from "./PaymentStatus";
+
+export type Payment = {
+  id: string,
+  dueDate: Date,
+  status: PaymentStatus,
+  category: string,
+  description: string,
+  amount: number
+}
